+++ google/cloud/bigtable/bigtable_client.bzl
@@ -37,7 +37,6 @@ bigtable_client_hdrs = [
     "instance_config.h",
     "instance_list_responses.h",
     "instance_update_config.h",
-    "internal/api_client_header.h",
     "internal/async_bulk_apply.h",
     "internal/async_longrunning_op.h",
     "internal/async_poll_op.h",
@@ -88,7 +87,6 @@ bigtable_client_srcs = [
     "instance_admin_client.cc",
     "instance_config.cc",
     "instance_update_config.cc",
-    "internal/api_client_header.cc",
     "internal/async_bulk_apply.cc",
     "internal/bulk_mutator.cc",
     "internal/common_client.cc",
