+++ google/cloud/bigtable/instance_admin_test.cc
@@ -13,10 +13,10 @@
 // limitations under the License.
 
 #include "google/cloud/bigtable/instance_admin.h"
-#include "google/cloud/bigtable/internal/api_client_header.h"
 #include "google/cloud/bigtable/testing/mock_async_failing_rpc_factory.h"
 #include "google/cloud/bigtable/testing/mock_instance_admin_client.h"
 #include "google/cloud/bigtable/testing/mock_response_reader.h"
+#include "google/cloud/internal/api_client_header.h"
 #include "google/cloud/testing_util/assert_ok.h"
 #include "google/cloud/testing_util/chrono_literals.h"
 #include "google/cloud/testing_util/mock_completion_queue.h"
@@ -63,7 +63,7 @@ auto create_list_instances_lambda =
         EXPECT_STATUS_OK(IsContextMDValid(
             *context,
             "google.bigtable.admin.v2.BigtableInstanceAdmin.ListInstances",
-            bigtable::internal::ApiClientHeader()));
+            google::cloud::internal::ApiClientHeader()));
         auto const project_name = "projects/" + kProjectId;
         EXPECT_EQ(project_name, request.parent());
         EXPECT_EQ(expected_token, request.page_token());
@@ -88,7 +88,7 @@ auto create_get_cluster_mock = []() {
             btadmin::Cluster* response) {
     EXPECT_STATUS_OK(IsContextMDValid(
         *context, "google.bigtable.admin.v2.BigtableInstanceAdmin.GetCluster",
-        bigtable::internal::ApiClientHeader()));
+        google::cloud::internal::ApiClientHeader()));
     EXPECT_NE(nullptr, response);
     response->set_name(request.name());
     return grpc::Status::OK;
@@ -101,7 +101,7 @@ auto create_get_policy_mock = []() {
             ::google::iam::v1::Policy* response) {
     EXPECT_STATUS_OK(IsContextMDValid(
         *context, "google.bigtable.admin.v2.BigtableInstanceAdmin.GetIamPolicy",
-        bigtable::internal::ApiClientHeader()));
+        google::cloud::internal::ApiClientHeader()));
     EXPECT_NE(nullptr, response);
     response->set_version(3);
     response->set_etag("random-tag");
@@ -115,7 +115,7 @@ auto create_policy_with_params = []() {
             ::google::iam::v1::Policy* response) {
     EXPECT_STATUS_OK(IsContextMDValid(
         *context, "google.bigtable.admin.v2.BigtableInstanceAdmin.SetIamPolicy",
-        bigtable::internal::ApiClientHeader()));
+        google::cloud::internal::ApiClientHeader()));
     EXPECT_NE(nullptr, response);
     *response = request.policy();
     return grpc::Status::OK;
@@ -135,7 +135,7 @@ auto create_list_clusters_lambda =
         EXPECT_STATUS_OK(IsContextMDValid(
             *context,
             "google.bigtable.admin.v2.BigtableInstanceAdmin.ListClusters",
-            bigtable::internal::ApiClientHeader()));
+            google::cloud::internal::ApiClientHeader()));
         auto const instance_name =
             "projects/" + kProjectId + "/instances/" + instance_id;
         EXPECT_EQ(instance_name, request.parent());
@@ -176,7 +176,7 @@ struct MockRpcFactory {
                                    RequestType const& request,
                                    ResponseType* response) {
           EXPECT_STATUS_OK(IsContextMDValid(
-              *context, method, bigtable::internal::ApiClientHeader()));
+              *context, method, google::cloud::internal::ApiClientHeader()));
           RequestType expected;
           // Cannot use ASSERT_TRUE() here, it has an embedded "return;"
           EXPECT_TRUE(google::protobuf::TextFormat::ParseFromString(
@@ -271,7 +271,7 @@ TEST_F(InstanceAdminTest, ListInstancesRecoverableFailures) {
     EXPECT_STATUS_OK(IsContextMDValid(
         *context,
         "google.bigtable.admin.v2.BigtableInstanceAdmin.ListInstances",
-        bigtable::internal::ApiClientHeader()));
+        google::cloud::internal::ApiClientHeader()));
     return grpc::Status(grpc::StatusCode::UNAVAILABLE, "try-again");
   };
   auto batch0 = create_list_instances_lambda("", "token-001", {"t0", "t1"});
@@ -371,7 +371,7 @@ TEST_F(InstanceAdminTest, ListClustersRecoverableFailures) {
                                      btadmin::ListClustersResponse*) {
     EXPECT_STATUS_OK(IsContextMDValid(
         *context, "google.bigtable.admin.v2.BigtableInstanceAdmin.ListClusters",
-        bigtable::internal::ApiClientHeader()));
+        google::cloud::internal::ApiClientHeader()));
     return grpc::Status(grpc::StatusCode::UNAVAILABLE, "try-again");
   };
   std::string const& instance_id = "the-instance";
@@ -442,7 +442,7 @@ TEST_F(InstanceAdminTest, GetClusterRecoverableError) {
                                      btadmin::Cluster*) {
     EXPECT_STATUS_OK(IsContextMDValid(
         *context, "google.bigtable.admin.v2.BigtableInstanceAdmin.GetCluster",
-        bigtable::internal::ApiClientHeader()));
+        google::cloud::internal::ApiClientHeader()));
     return grpc::Status(grpc::StatusCode::UNAVAILABLE, "try-again");
   };
 
@@ -519,7 +519,7 @@ TEST_F(InstanceAdminTest, GetIamPolicyWithConditionsFails) {
         EXPECT_STATUS_OK(IsContextMDValid(
             *context,
             "google.bigtable.admin.v2.BigtableInstanceAdmin.GetIamPolicy",
-            bigtable::internal::ApiClientHeader()));
+            google::cloud::internal::ApiClientHeader()));
         EXPECT_NE(nullptr, response);
         response->set_version(3);
         response->set_etag("random-tag");
@@ -566,7 +566,7 @@ TEST_F(InstanceAdminTest, GetIamPolicyRecoverableError) {
                                      iamproto::Policy*) {
     EXPECT_STATUS_OK(IsContextMDValid(
         *context, "google.bigtable.admin.v2.BigtableInstanceAdmin.GetIamPolicy",
-        bigtable::internal::ApiClientHeader()));
+        google::cloud::internal::ApiClientHeader()));
     return grpc::Status(grpc::StatusCode::UNAVAILABLE, "try-again");
   };
   auto mock_policy = create_get_policy_mock();
@@ -622,7 +622,7 @@ TEST_F(InstanceAdminTest, GetNativeIamPolicyRecoverableError) {
                                      iamproto::Policy*) {
     EXPECT_STATUS_OK(IsContextMDValid(
         *context, "google.bigtable.admin.v2.BigtableInstanceAdmin.GetIamPolicy",
-        bigtable::internal::ApiClientHeader()));
+        google::cloud::internal::ApiClientHeader()));
     return grpc::Status(grpc::StatusCode::UNAVAILABLE, "try-again");
   };
   auto mock_policy = create_get_policy_mock();
@@ -657,7 +657,7 @@ class AsyncGetIamPolicyTest : public ::testing::Test {
           EXPECT_STATUS_OK(IsContextMDValid(
               *context,
               "google.bigtable.admin.v2.BigtableInstanceAdmin.GetIamPolicy",
-              bigtable::internal::ApiClientHeader()));
+              google::cloud::internal::ApiClientHeader()));
           EXPECT_EQ("projects/the-project/instances/test-instance",
                     request.resource());
           // This is safe, see comments in MockAsyncResponseReader.
@@ -828,7 +828,7 @@ TEST_F(InstanceAdminTest, SetIamPolicyRecoverableError) {
                                      iamproto::Policy*) {
     EXPECT_STATUS_OK(IsContextMDValid(
         *context, "google.bigtable.admin.v2.BigtableInstanceAdmin.SetIamPolicy",
-        bigtable::internal::ApiClientHeader()));
+        google::cloud::internal::ApiClientHeader()));
     return grpc::Status(grpc::StatusCode::UNAVAILABLE, "try-again");
   };
   auto mock_policy = create_policy_with_params();
@@ -896,7 +896,7 @@ TEST_F(InstanceAdminTest, SetNativeIamPolicyRecoverableError) {
                                      iamproto::Policy*) {
     EXPECT_STATUS_OK(IsContextMDValid(
         *context, "google.bigtable.admin.v2.BigtableInstanceAdmin.SetIamPolicy",
-        bigtable::internal::ApiClientHeader()));
+        google::cloud::internal::ApiClientHeader()));
     return grpc::Status(grpc::StatusCode::UNAVAILABLE, "try-again");
   };
   auto mock_policy = create_policy_with_params();
@@ -928,7 +928,7 @@ TEST_F(InstanceAdminTest, TestIamPermissions) {
         EXPECT_STATUS_OK(IsContextMDValid(
             *context,
             "google.bigtable.admin.v2.BigtableInstanceAdmin.TestIamPermissions",
-            bigtable::internal::ApiClientHeader()));
+            google::cloud::internal::ApiClientHeader()));
         EXPECT_NE(nullptr, response);
         std::vector<std::string> permissions = {"writer", "reader"};
         response->add_permissions(permissions[0]);
@@ -976,7 +976,7 @@ TEST_F(InstanceAdminTest, TestIamPermissionsRecoverableError) {
     EXPECT_STATUS_OK(IsContextMDValid(
         *context,
         "google.bigtable.admin.v2.BigtableInstanceAdmin.TestIamPermissions",
-        bigtable::internal::ApiClientHeader()));
+        google::cloud::internal::ApiClientHeader()));
     return grpc::Status(grpc::StatusCode::UNAVAILABLE, "try-again");
   };
 
@@ -987,7 +987,7 @@ TEST_F(InstanceAdminTest, TestIamPermissionsRecoverableError) {
         EXPECT_STATUS_OK(IsContextMDValid(
             *context,
             "google.bigtable.admin.v2.BigtableInstanceAdmin.TestIamPermissions",
-            bigtable::internal::ApiClientHeader()));
+            google::cloud::internal::ApiClientHeader()));
         EXPECT_NE(nullptr, response);
         std::vector<std::string> permissions = {"writer", "reader"};
         response->add_permissions(permissions[0]);
@@ -1025,7 +1025,7 @@ class AsyncDeleteClusterTest : public ::testing::Test {
           EXPECT_STATUS_OK(IsContextMDValid(
               *context,
               "google.bigtable.admin.v2.BigtableInstanceAdmin.DeleteCluster",
-              bigtable::internal::ApiClientHeader()));
+              google::cloud::internal::ApiClientHeader()));
           EXPECT_EQ(
               "projects/the-project/instances/test-instance/clusters/"
               "the-cluster",
@@ -1110,7 +1110,7 @@ class AsyncSetIamPolicyTest : public ::testing::Test {
           EXPECT_STATUS_OK(IsContextMDValid(
               *context,
               "google.bigtable.admin.v2.BigtableInstanceAdmin.SetIamPolicy",
-              bigtable::internal::ApiClientHeader()));
+              google::cloud::internal::ApiClientHeader()));
           EXPECT_EQ("projects/the-project/instances/test-instance",
                     request.resource());
           // This is safe, see comments in MockAsyncResponseReader.
@@ -1273,7 +1273,7 @@ class AsyncTestIamPermissionsTest : public ::testing::Test {
                   *context,
                   "google.bigtable.admin.v2.BigtableInstanceAdmin."
                   "TestIamPermissions",
-                  bigtable::internal::ApiClientHeader()));
+                  google::cloud::internal::ApiClientHeader()));
               EXPECT_EQ("projects/the-project/instances/the-resource",
                         request.resource());
               // This is safe, see comments in MockAsyncResponseReader.
