+++ google/cloud/bigtable/mutation_batcher_test.cc
@@ -13,10 +13,10 @@
 // limitations under the License.
 
 #include "google/cloud/bigtable/mutation_batcher.h"
-#include "google/cloud/bigtable/internal/api_client_header.h"
 #include "google/cloud/bigtable/testing/mock_mutate_rows_reader.h"
 #include "google/cloud/bigtable/testing/table_test_fixture.h"
 #include "google/cloud/future.h"
+#include "google/cloud/internal/api_client_header.h"
 #include "google/cloud/testing_util/assert_ok.h"
 #include "google/cloud/testing_util/chrono_literals.h"
 #include "google/cloud/testing_util/mock_completion_queue.h"
@@ -170,7 +170,7 @@ class MutationBatcherTest : public bigtable::testing::TableTestFixture {
                                        grpc::CompletionQueue*) {
             EXPECT_STATUS_OK(IsContextMDValid(
                 *context, "google.bigtable.v2.Bigtable.MutateRows",
-                bigtable::internal::ApiClientHeader()));
+                google::cloud::internal::ApiClientHeader()));
             EXPECT_EQ(exchange.req.size(), r.entries_size());
             for (std::size_t i = 0; i != exchange.req.size(); ++i) {
               google::bigtable::v2::MutateRowsRequest::Entry expected;
