+++ google/cloud/bigtable/testing/mock_async_failing_rpc_factory.h
@@ -15,8 +15,8 @@
 #ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_BIGTABLE_TESTING_MOCK_ASYNC_FAILING_RPC_FACTORY_H
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_BIGTABLE_TESTING_MOCK_ASYNC_FAILING_RPC_FACTORY_H
 
-#include "google/cloud/bigtable/internal/api_client_header.h"
 #include "google/cloud/bigtable/testing/mock_response_reader.h"
+#include "google/cloud/internal/api_client_header.h"
 #include "google/cloud/status.h"
 #include "google/cloud/testing_util/validate_metadata.h"
 #include <google/protobuf/text_format.h>
@@ -58,7 +58,7 @@ struct MockAsyncFailingRpcFactory {
                                             grpc::CompletionQueue*) {
       using ::testing::_;
       EXPECT_STATUS_OK(google::cloud::testing_util::IsContextMDValid(
-          *context, method, bigtable::internal::ApiClientHeader()));
+          *context, method, google::cloud::internal::ApiClientHeader()));
       RequestType expected;
       // Cannot use ASSERT_TRUE() here, it has an embedded "return;"
       EXPECT_TRUE(google::protobuf::TextFormat::ParseFromString(
