+++ google/cloud/bigtable/testing/mock_response_reader.h
@@ -15,7 +15,7 @@
 #ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_BIGTABLE_TESTING_MOCK_RESPONSE_READER_H
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_BIGTABLE_TESTING_MOCK_RESPONSE_READER_H
 
-#include "google/cloud/bigtable/internal/api_client_header.h"
+#include "google/cloud/internal/api_client_header.h"
 #include "google/cloud/testing_util/assert_ok.h"
 #include "google/cloud/testing_util/validate_metadata.h"
 #include <gmock/gmock.h>
@@ -67,7 +67,7 @@ class MockResponseReader : public grpc::ClientReaderInterface<Response> {
   MakeMockReturner() {
     return [this](grpc::ClientContext* context, Request const&) {
       EXPECT_STATUS_OK(google::cloud::testing_util::IsContextMDValid(
-          *context, method_, internal::ApiClientHeader()));
+          *context, method_, google::cloud::internal::ApiClientHeader()));
       return UniquePtr(this);
     };
   }
