+++ google/cloud/google_cloud_cpp_common_unit_tests.bzl
@@ -22,6 +22,7 @@ google_cloud_cpp_common_unit_tests = [
     "future_void_test.cc",
     "future_void_then_test.cc",
     "iam_bindings_test.cc",
+    "internal/api_client_header_test.cc",
     "internal/backoff_policy_test.cc",
     "internal/big_endian_test.cc",
     "internal/compiler_info_test.cc",
