+++ google/cloud/internal/api_client_header.h
@@ -1,4 +1,4 @@
-// Copyright 2020 Google LLC
+// Copyright 2019 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,26 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/bigtable/internal/api_client_header.h"
-#include "google/cloud/internal/compiler_info.h"
+#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_API_CLIENT_HEADER_H
+#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_API_CLIENT_HEADER_H
+
+#include "google/cloud/version.h"
 #include <string>
 
 namespace google {
 namespace cloud {
-namespace bigtable {
-inline namespace BIGTABLE_CLIENT_NS {
+inline namespace GOOGLE_CLOUD_CPP_NS {
 namespace internal {
 
-std::string ApiClientHeader() {
-  return "gl-cpp/" + google::cloud::internal::CompilerId() + "-" +
-         google::cloud::internal::CompilerVersion() + "-" +
-         google::cloud::internal::CompilerFeatures() + "-" +
-         google::cloud::internal::LanguageVersion() + " gccl/" +
-         version_string();
-}
+/// Return the value for the x-goog-api-client header (aka metadata).
+std::string ApiClientHeader();
 
 }  // namespace internal
-}  // namespace BIGTABLE_CLIENT_NS
-}  // namespace bigtable
+}  // namespace GOOGLE_CLOUD_CPP_NS
 }  // namespace cloud
 }  // namespace google
+
+#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_API_CLIENT_HEADER_H
