+++ google/cloud/spanner/CMakeLists.txt
@@ -66,8 +66,6 @@ add_library(
     instance_admin_client.h
     instance_admin_connection.cc
     instance_admin_connection.h
-    internal/api_client_header.cc
-    internal/api_client_header.h
     internal/channel.h
     internal/clock.h
     internal/connection_impl.cc
@@ -259,7 +257,6 @@ function (spanner_client_define_tests)
         instance_admin_client_test.cc
         instance_admin_connection_test.cc
         instance_test.cc
-        internal/api_client_header_test.cc
         internal/clock_test.cc
         internal/connection_impl_test.cc
         internal/database_admin_logging_test.cc
