--- google/cloud/spanner/internal/api_client_header.h
@@ -1,36 +1,0 @@
-// Copyright 2019 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_SPANNER_INTERNAL_API_CLIENT_HEADER_H
-#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_SPANNER_INTERNAL_API_CLIENT_HEADER_H
-
-#include "google/cloud/spanner/version.h"
-#include <string>
-
-namespace google {
-namespace cloud {
-namespace spanner {
-inline namespace SPANNER_CLIENT_NS {
-namespace internal {
-
-/// Return the value for the x-goog-api-client header (aka metadata).
-std::string ApiClientHeader();
-
-}  // namespace internal
-}  // namespace SPANNER_CLIENT_NS
-}  // namespace spanner
-}  // namespace cloud
-}  // namespace google
-
-#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_SPANNER_INTERNAL_API_CLIENT_HEADER_H
