+++ google/cloud/spanner/internal/metadata_spanner_stub_test.cc
@@ -14,8 +14,8 @@
 
 #include "google/cloud/spanner/internal/metadata_spanner_stub.h"
 #include "google/cloud/spanner/database.h"
-#include "google/cloud/spanner/internal/api_client_header.h"
 #include "google/cloud/spanner/testing/mock_spanner_stub.h"
+#include "google/cloud/internal/api_client_header.h"
 #include "google/cloud/testing_util/assert_ok.h"
 #include "google/cloud/testing_util/validate_metadata.h"
 #include <gmock/gmock.h>
@@ -40,7 +40,7 @@ class MetadataSpannerStubTest : public ::testing::Test {
  protected:
   void SetUp() override {
     mock_ = std::make_shared<spanner_testing::MockSpannerStub>();
-    expected_api_client_header_ = ApiClientHeader();
+    expected_api_client_header_ = google::cloud::internal::ApiClientHeader();
   }
 
   void TearDown() override {}
