+++ google/cloud/spanner/spanner_client.bzl
@@ -35,7 +35,6 @@ spanner_client_hdrs = [
     "instance.h",
     "instance_admin_client.h",
     "instance_admin_connection.h",
-    "internal/api_client_header.h",
     "internal/channel.h",
     "internal/clock.h",
     "internal/connection_impl.h",
@@ -94,7 +93,6 @@ spanner_client_srcs = [
     "instance.cc",
     "instance_admin_client.cc",
     "instance_admin_connection.cc",
-    "internal/api_client_header.cc",
     "internal/connection_impl.cc",
     "internal/database_admin_logging.cc",
     "internal/database_admin_metadata.cc",
