+++ google/cloud/storage/version.cc
@@ -13,10 +13,8 @@
 // limitations under the License.
 
 #include "google/cloud/storage/version.h"
-#include "google/cloud/internal/build_info.h"
-#include "google/cloud/internal/compiler_info.h"
+#include "google/cloud/internal/api_client_header.h"
 #include <limits>
-#include <sstream>
 
 namespace google {
 namespace cloud {
@@ -25,16 +23,7 @@ inline namespace STORAGE_CLIENT_NS {
 std::string version_string() { return ::google::cloud::version_string(); }
 
 std::string x_goog_api_client() {
-  static std::string const kXGoogApiClient = [] {
-    std::ostringstream os;
-    os << "gl-cpp/" << google::cloud::internal::CompilerId() << "-"
-       << google::cloud::internal::CompilerVersion() << "-"
-       << google::cloud::internal::CompilerFeatures() << "-"
-       << google::cloud::internal::LanguageVersion() << " gccl/"
-       << version_string();
-    return os.str();
-  }();
-  return kXGoogApiClient;
+  return google::cloud::internal::ApiClientHeader();
 }
 
 // These were sprinkled through the code, consolidated here because I could
